airflow-docker/requirements.txt
# FinanceDataProject/airflow/requirements.txt
redis
psycopg2-binary
apache-airflow[cassandra,celery,postgres,redis]
selenium
webdriver-manager
cassandra-driver
pandas_market_calendars
lxml
polars
bs4
yfinance
tabulate
pylint
openpyxl
schedule
pyspark
pandas
numpy
scipy
timeout-decorator
matplotlib
# 